api_clientes/requirements.txt
flask
mysql-connector-python 